Renamed the packages to tipsymcstagger.com Added configuration for publishing to maven central
