+++ src/test/java/com/tipsymcstagger/spring/request/correlation/demo/DemoConfiguration.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2015-2024 the original author or authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");  you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied.  See the License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package com.tipsymcstagger.spring.request.correlation.demo;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+/**
+ * Configuration class for the Demo application. It declares the Rest Template bean because we can't
+ * declare it and use it in the same class.
+ *
+ * @author Steven C. Saliman
+ */
+@Configuration
+public class DemoConfiguration {
+  @Bean
+  public RestTemplate restTemplate() {
+    return new RestTemplate();
+  }
+}
