+++ src/test/java/com/tipsymcstagger/spring/request/correlation/demo/DemoFeignClient.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2015-2024 the original author or authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");  you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied.  See the License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package com.tipsymcstagger.spring.request.correlation.demo;
+
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+/**
+ * A Feign client we can use to prove that correlating ids propagate to it.
+ *
+ * @author Steven C. Saliman
+ */
+@FeignClient(name = "local", url = "localhost:10344")
+interface DemoFeignClient {
+
+  @RequestMapping(value = "/ids", method = RequestMethod.GET)
+  String getCorrelatingIds();
+}
